refactor(runtime): the adversarial battery says its name (tests.rs)
The #[cfg(test)]-gated adversarial module (mod.rs docstring: "Test-only:
no public surface, no production code") was spread over three files the
15k prod counter reads as production — 1093 LOC of test machinery
charged against the maintainability budget, tripping vector 24 at the
lot-3a push (15173 > 15000). The counter's own convention names the
fix: a test module split into its own file is called tests.rs
(check-crate-size.sh · the rs_prod_files rule). The threat-model doc
stays in mod.rs where a reader lands; the fixture plane and the
assertion harness fold into adversarial/tests.rs as inner modules —
super::fixtures paths hold by construction, zero semantic change.

nika-runtime prod LOC: 15173 to 14187. Battery green (f1..f5 + the
suite contract). fmt clean · vector 24 OK.
