+++ crates/nika-runtime/src/adversarial/mod.rs
@@ -105,173 +105,5 @@
 //! this suite pins the deterministic guarantees the engine makes
 //! regardless of which model is plugged in.
 
-#![allow(clippy::expect_used, clippy::unwrap_used, clippy::panic)]
-
-mod fixtures;
-mod harness;
-
-use fixtures::{Containment, ExpectedStatus, Fixture, Verdict};
-
-/// Execute one fixture under its declared verdict.
-async fn execute(fx: &Fixture) {
-    match fx.sidecar.verdict {
-        Verdict::StaticDeny => harness::assert_static_deny(fx),
-        Verdict::RuntimeBlock => harness::assert_runtime(fx).await,
-        Verdict::Residual => {
-            let residual = fx.sidecar.residual.as_ref().unwrap_or_else(|| {
-                panic!(
-                    "{}: verdict residual REQUIRES a residual block — a \
-                     documented gap is a TODO with an owner, never a silent \
-                     pass",
-                    fx.id()
-                )
-            });
-            assert!(
-                !residual.summary.is_empty() && !residual.owner.is_empty(),
-                "{}: the residual note names the gap AND its owner",
-                fx.id()
-            );
-            harness::assert_runtime(fx).await;
-        }
-    }
-}
-
-/// Run every fixture of one family directory.
-async fn run_family(family_dir: &str) {
-    let mut ran = 0usize;
-    for fx in fixtures::load_all() {
-        if fx.family_dir == family_dir {
-            execute(&fx).await;
-            ran += 1;
-        }
-    }
-    assert!(
-        ran >= 3,
-        "{family_dir}: each family carries at least 3 fixtures"
-    );
-}
-
-#[tokio::test]
-async fn f1_direct_injection() {
-    run_family("f1-direct-injection").await;
-}
-
-#[tokio::test]
-async fn f2_indirect_tool_output() {
-    run_family("f2-indirect-tool-output").await;
-}
-
-#[tokio::test]
-async fn f3_schema_parse_smuggling() {
-    run_family("f3-schema-parse-smuggling").await;
-}
-
-#[tokio::test]
-async fn f4_permit_escalation() {
-    run_family("f4-permit-escalation").await;
-}
-
-#[tokio::test]
-async fn f5_multi_hop_laundering() {
-    run_family("f5-multi-hop-laundering").await;
-}
-
-/// The suite's own contract: sidecars are well-formed, verdicts pair with
-/// their containment class, every family is covered, and the residual
-/// registry never silently empties.
-#[test]
-fn suite_contract() {
-    let all = fixtures::load_all();
-    assert!(
-        all.len() >= 15,
-        "the suite covers at least 15 attack fixtures — found {}",
-        all.len()
-    );
-    let mut per_family = [0usize; 5];
-    let mut residuals = 0usize;
-    for (i, (tag, dir)) in fixtures::FAMILIES.iter().enumerate() {
-        for fx in &all {
-            if fx.family_dir == *dir {
-                per_family[i] += 1;
-                check_sidecar(fx, tag);
-            }
-        }
-        assert!(
-            per_family[i] >= 3,
-            "{dir}: at least 3 fixtures per family — found {}",
-            per_family[i]
-        );
-    }
-    for fx in &all {
-        if fx.sidecar.verdict == Verdict::Residual {
-            residuals += 1;
-        }
-    }
-    assert!(
-        residuals >= 1,
-        "the residual registry is empty — if every gap closed, update the \
-         suite doc's residual section deliberately"
-    );
-}
-
-/// One sidecar's internal consistency (family tag, slug, verdict pairing).
-fn check_sidecar(fx: &Fixture, want_family: &str) {
-    let sc = &fx.sidecar;
-    assert_eq!(
-        sc.family,
-        want_family,
-        "{}: sidecar family matches its directory",
-        fx.id()
-    );
-    assert!(
-        !sc.title.is_empty() && !sc.rationale.is_empty(),
-        "{}: title and rationale are non-empty",
-        fx.id()
-    );
-    match sc.verdict {
-        Verdict::StaticDeny => {
-            assert!(
-                matches!(
-                    sc.containment,
-                    Containment::TrifectaStatic | Containment::PermitsStatic
-                ),
-                "{}: static-deny rides a static containment class",
-                fx.id()
-            );
-            assert!(
-                sc.static_expect.is_some() && sc.script.is_none() && sc.expect.is_none(),
-                "{}: static-deny carries `static` and no runtime blocks",
-                fx.id()
-            );
-        }
-        Verdict::RuntimeBlock => {
-            assert!(
-                matches!(
-                    sc.containment,
-                    Containment::AgentWhitelist
-                        | Containment::ExecPermitGate
-                        | Containment::FsBoundary
-                ),
-                "{}: runtime-block rides a runtime containment class",
-                fx.id()
-            );
-            let expect = sc
-                .expect
-                .as_ref()
-                .unwrap_or_else(|| panic!("{}: runtime-block requires `expect`", fx.id()));
-            assert!(
-                expect.status != Some(ExpectedStatus::Success) && expect.code.is_some(),
-                "{}: a block pins a failure code",
-                fx.id()
-            );
-        }
-        Verdict::Residual => {
-            assert_eq!(
-                sc.containment,
-                Containment::None,
-                "{}: a residual is containment: none (honesty — never fake a block)",
-                fx.id()
-            );
-        }
-    }
-}
+#[cfg(test)]
+mod tests;
